+++ CHANGELOG.md
@@ -22,6 +22,8 @@ Changes since v2.15
 
 - The OIDC configuration is now validated, and REMS refuses to start without a valid OIDC configuration. See the `:oidc-metadata-url` configuration variable and [configuration.md](docs/configuration.md). (#2519)
 
+- The navigation menu brand logo is now configurable. You can see `:theme :navbar-logo` configuration variable that if set to true adds logo top navigation menu. [configuration.md](docs/configuration.md). (#2363)
+
 ## v2.15 "Tapiolantie" 2021-01-08
 
 ### Changes